make it easier to use in different ways
